handshake: library polish
